Upgrade and fix preview-link code sample for Python to enable create 3P resources
