+++ python/3p-resources/create_3p_resources/requirements.txt
@@ -1,0 +1,2 @@
+Flask>=2.2.2
+functions-framework>=3.5.0
